古いAPIにObsolute属性を付けた
