+++ src/BeUtl.Graphics/Graphics/NullCanvas.cs
@@ -58,9 +58,10 @@ internal sealed class NullCanvas : ICanvas
 
     public void DrawRect(Size size) => throw new NotImplementedException();
 
-    public void DrawText(FormattedText_ text) => throw new NotImplementedException();
+    public void DrawText(FormattedText text) => throw new NotImplementedException();
 
-    public void DrawText(TextElement text, Size size) => throw new NotImplementedException();
+    [Obsolete("Use 'DrawText(FormattedText)'.")]
+    public void DrawText(Media.TextFormatting.Compat.TextElement text, Size size) => throw new NotImplementedException();
 
     public void FillCircle(Size size) => throw new NotImplementedException();
