+++ src/BeUtl.Graphics/Media/TextFormatting/FormattedText.cs
@@ -1,5 +1,4 @@
 ﻿using System.Diagnostics;
-using System.Text.Json.Nodes;
 
 using BeUtl.Graphics;
 
@@ -8,7 +7,7 @@ using SkiaSharp;
 namespace BeUtl.Media.TextFormatting;
 
 [DebuggerDisplay("{Text}")]
-public struct FormattedText_ : IEquatable<FormattedText_>
+public struct FormattedText : IEquatable<FormattedText>
 {
     private FontWeight _weight = FontWeight.Regular;
     private FontStyle _style = FontStyle.Normal;
@@ -20,7 +19,7 @@ public struct FormattedText_ : IEquatable<FormattedText_>
     private Size _bounds = default;
     private bool _isDirty = false;
 
-    public FormattedText_()
+    public FormattedText()
     {
     }
 
@@ -120,17 +119,17 @@ public struct FormattedText_ : IEquatable<FormattedText_>
 
     public override bool Equals(object? obj)
     {
-        return obj is FormattedText_ text && Equals(text);
+        return obj is FormattedText text && Equals(text);
     }
 
-    public bool Equals(FormattedText_ other)
+    public bool Equals(FormattedText other)
     {
         return Weight == other.Weight && Style == other.Style && Font.Equals(other.Font) && Size == other.Size && Spacing == other.Spacing && Text.Equals(other.Text) && BeginOnNewLine == other.BeginOnNewLine && EqualityComparer<IBrush>.Default.Equals(Brush, other.Brush) && Margin.Equals(other.Margin);
     }
 
     public override int GetHashCode()
     {
-        HashCode hash = new HashCode();
+        HashCode hash = new();
         hash.Add(Weight);
         hash.Add(Style);
         hash.Add(Font);
@@ -143,147 +142,7 @@ public struct FormattedText_ : IEquatable<FormattedText_>
         return hash.ToHashCode();
     }
 
-    public static bool operator ==(FormattedText_ left, FormattedText_ right) => left.Equals(right);
+    public static bool operator ==(FormattedText left, FormattedText right) => left.Equals(right);
 
-    public static bool operator !=(FormattedText_ left, FormattedText_ right) => !(left == right);
-}
-
-public sealed class TextLines : AffectsRenders<TextLine>
-{
-
-}
-
-public class FormattedText : Drawable
-{
-    public static readonly CoreProperty<TextLines> LinesProperty;
-    private readonly TextLines _lines;
-
-    static FormattedText()
-    {
-        LinesProperty = ConfigureProperty<TextLines, FormattedText>(nameof(Lines))
-            .Accessor(o => o.Lines, (o, v) => o.Lines = v)
-            .PropertyFlags(PropertyFlags.KnownFlags_1)
-            .Register();
-    }
-
-    public FormattedText()
-    {
-        _lines = new()
-        {
-            Attached = item => (item as ILogicalElement).NotifyAttachedToLogicalTree(new(this)),
-            Detached = item => (item as ILogicalElement).NotifyDetachedFromLogicalTree(new(this)),
-        };
-        _lines.Invalidated += (_, _) => Invalidate();
-    }
-
-    private FormattedText(List<TextLine> lines)
-        : this()
-    {
-        _lines.AddRange(lines);
-    }
-
-    public TextLines Lines
-    {
-        get => _lines;
-        set => _lines.Replace(value);
-    }
-
-    public static FormattedText Parse(string s, FormattedTextInfo info)
-    {
-        var tokenizer = new FormattedTextParser(s);
-        List<TextLine> lines = tokenizer.ToLines(info);
-
-        return new FormattedText(lines);
-    }
-
-    public void Load(string s, FormattedTextInfo info)
-    {
-        var tokenizer = new FormattedTextParser(s);
-        List<TextLine> lines = tokenizer.ToLines(info);
-
-        _lines.Replace(lines);
-        Invalidate();
-    }
-
-    protected override void OnDraw(ICanvas canvas)
-    {
-        DrawCore(canvas);
-    }
-
-    private void DrawCore(ICanvas canvas)
-    {
-        float prevBottom = 0;
-        for (int i = 0; i < Lines.Count; i++)
-        {
-            TextLine line = Lines[i];
-            line.Measure(canvas.Size.ToSize(1));
-            Rect lineBounds = line.Bounds;
-
-            using (canvas.PushTransform(Matrix.CreateTranslation(0, prevBottom)))
-            {
-                line.Draw(canvas);
-                prevBottom += lineBounds.Height;
-            }
-        }
-    }
-
-    protected override Size MeasureCore(Size availableSize)
-    {
-        float width = 0;
-        float height = 0;
-
-        foreach (TextLine line in Lines)
-        {
-            line.Measure(availableSize);
-            Rect bounds = line.Bounds;
-            width = MathF.Max(bounds.Width, width);
-            height += bounds.Height;
-        }
-
-        return new Size(width, height);
-    }
-
-    public override void ReadFromJson(JsonNode json)
-    {
-        base.ReadFromJson(json);
-        if (json is JsonObject jobject)
-        {
-            if (jobject.TryGetPropertyValue("lines", out JsonNode? childrenNode)
-                && childrenNode is JsonArray childrenArray)
-            {
-                _lines.Clear();
-                if (_lines.Capacity < childrenArray.Count)
-                {
-                    _lines.Capacity = childrenArray.Count;
-                }
-
-                foreach (JsonObject childJson in childrenArray.OfType<JsonObject>())
-                {
-                    var item = new TextLine();
-                    item.ReadFromJson(childJson);
-                    _lines.Add(item);
-                }
-            }
-        }
-    }
-
-    public override void WriteToJson(ref JsonNode json)
-    {
-        base.WriteToJson(ref json);
-
-        if (json is JsonObject jobject)
-        {
-            var array = new JsonArray();
-
-            foreach (TextLine item in _lines.AsSpan())
-            {
-                JsonNode node = new JsonObject();
-                item.WriteToJson(ref node);
-
-                array.Add(node);
-            }
-
-            jobject["lines"] = array;
-        }
-    }
+    public static bool operator !=(FormattedText left, FormattedText right) => !(left == right);
 }
