+++ tests/BeUtl.Benchmarks/BeUtl.Benchmarks.csproj
@@ -1,4 +1,4 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
+<Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
     <OutputType>Exe</OutputType>
@@ -20,5 +20,4 @@
     <ProjectReference Include="..\..\src\BeUtl.Operations\BeUtl.Operations.csproj" />
     <ProjectReference Include="..\..\src\BeUtl.ProjectSystem\BeUtl.ProjectSystem.csproj" />
   </ItemGroup>
-
 </Project>
